fix task manager http server may be blocked bug and add two service script
