+++ arch/conf/server_conf.json
@@ -23,6 +23,5 @@
 			"127.0.0.1:8001"
 		]
 	},
-	"party_id": 10000,
-  	"all_party_ids": [9999, 10000]
+	"party_id": 10000
 }
