+++ arch/fate-serving/serving/src/main/resources/serving-server.properties
@@ -1,0 +1,38 @@
+#
+# Copyright 2019 The FATE Authors. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, 
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+coordinator=webank
+ip=127.0.0.1
+port=8001
+workMode=0
+
+#storage
+# maybe python/data/
+standaloneStoragePath=
+modelCacheAccessTTL=12
+modelCacheMaxSize=50
+
+# federation
+partyId=10000
+
+# adapter
+OnlineDataAccessAdapter=TestFile
+#OnlineDataAccessAdapter=DTest
+InferenceResultProcessingAdapter=PassProcessing
+
+# other subsystem
+proxy=127.0.0.1:9370
+roll=127.0.0.1:8011
