--- arch/task_manager/apps/server_status.py
@@ -1,26 +1,0 @@
-#
-#  Copyright 2019 The FATE Authors. All Rights Reserved.
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
-#
-from flask import Flask
-from arch.task_manager.utils.job_utils import get_json_result
-from arch.task_manager.settings import logger
-
-manager = Flask(__name__)
-
-
-@manager.route('/federated/heartbeat', methods=['POST'])
-def federated_heartbeat():
-    return get_json_result()
-
