+++ cluster-deploy/example-dir-tree/fate-serving/conf/serving-server.properties
@@ -21,7 +21,7 @@ workMode=0
 
 #storage
 # maybe python/data/
-standaloneStoragePath=/Users/jarviszeng/Work/Project/FDN/FATE/data/
+standaloneStoragePath=
 modelCacheAccessTTL=12
 modelCacheMaxSize=50
